Beta second deploy (#230)
* feat: missed a state in SP, some comments

* feat: updated tomb build

* fix: remove random debug statement

* fix: changed types for key management (#163)

* feat: fresh build of tomb from main

* feat: start deploying to beta

* chore: clean up service names

* feat: tweaks to deployment configs

* feat: unecessary feature

* feat: fight the deployment

* fix: frontend service name

* feat: need to use the entire directory now for the build unfortunately, guess that give better cacheing though

* fix: whoops left in a bit of a merge conflict

* fix: also fix the staging service

* chore: fresh wasm build, straight from main

* feat: just do automatic key generation

* fix: merge conflict

* fix: error types

* fix: duplicate error use

* fix: weird rebuild always happening in staging

* feat: better error handling on env reset

* feat: allow dry runs of build and deploy script

* fix: reset environment with key paths{

* fix: also deploy the built frontend

* feat: add dry run to staging deploy

* feat: restore parent docker building

* fix: staging service is now deployable again

* fix: storage provider isn't auto-deployed

* fix: rsync

* feat: remove file that slipped in from a merge conflict

* fix: clear out dist directory before building to limit what gets deployed

* fix: keep important files, ensure deps exist for the frontend

* fix: don't try to override deployed attributes and permissions

* fix: deployment issues

* fix: include dist

* fix: port mapping

* feat: try to work around localhost hardcoding

* fix: BAD BAD hardcoded localhost breaking production

* feat: fix service name in staging

* chore: update key names to match the other services

* chore: fresh build of banyan-cli

* chore: clean up, move yarn build before container

* feat: checkpoint for branch switching, test harnesses for upload reporting

* feat: test harnesses, and coverage over some metadata

* feat: reorganization and simplification

* fix: protect the active metadata

* fix: revert the hardcoded variables so this branch can be merged

* feat: update the wasm package to the newest main build

* fix: use correct verification key

* fix: lints

* fix: clippy lints

* feat: fix name in .env.sample

---------

+++ bin/reset_env.sh
@@ -5,22 +5,25 @@ set -o pipefail
 set -o nounset
 
 cd $(pwd)
-pwd
 
-# Remove all data
+function fail() {
+	local status_code=${1:-1}
+	local err_msg="${2:-an unknown error occurred}"
+
+	echo "ERROR: ${err_msg}"
+	exit $status_code
+}
+
 rm -rf \
-	crates/banyan-core-service/data/service-key* \
-	crates/banyan-core-service/data/server* \
-	crates/banyan-core-service/data/uploads/* \
-	\
-	crates/banyan-staging-service/data/service-key* \
-	crates/banyan-staging-service/data/platform-key* \
-	crates/banyan-staging-service/data/server* \
-	crates/banyan-staging-service/data/uploads/* \
-	\
-	crates/banyan-storage-provider-service/data/service-key* \
-	crates/banyan-storage-provider-service/data/platform-key* \
-	crates/banyan-storage-provider-service/data/server* \
+	crates/banyan-core-service/data/serv* \
+	crates/banyan-core-service/data/uploads/*
+
+rm -rf crates/banyan-staging-service/data/serv* \
+	crates/banyan-staging-service/data/platform* \
+	crates/banyan-staging-service/data/uploads/*
+
+rm -rf crates/banyan-storage-provider-service/data/serv* \
+	crates/banyan-storage-provider-service/data/platform* \
 	crates/banyan-storage-provider-service/data/uploads/*
 
 # Run core
@@ -30,6 +33,8 @@ rm -rf \
 	timeout 3s cargo run || true
 )
 
+[ -f "crates/banyan-core-service/data/service-key.public" ] || fail 1 "core didn't generate public key"
+
 # Copy core public key to staging and storage provider
 cp -f crates/banyan-core-service/data/service-key.public crates/banyan-staging-service/data/platform-key.public
 cp -f crates/banyan-core-service/data/service-key.public crates/banyan-storage-provider-service/data/platform-key.public
@@ -38,8 +43,23 @@ cp -f crates/banyan-core-service/data/service-key.public crates/banyan-storage-p
 (
 	cd crates/banyan-staging-service
 	cargo build
-	timeout 3s cargo run || true 
+	timeout 3s cargo run || true
 )
+
+[ -f "crates/banyan-staging-service/data/service-key.public" ] || fail 2 "staging missing public service key"
+[ -f "crates/banyan-staging-service/data/service-key.fingerprint" ] || fail 3 "staging missing service fingerprint"
+
+export STAGING_HOST_PUBKEY="$(cat crates/banyan-staging-service/data/service-key.public)"
+export STAGING_HOST_FINGERPRINT="$(cat crates/banyan-staging-service/data/service-key.fingerprint)"
+export STAGING_HOST_NAME="banyan-staging"
+export STAGING_HOST_URL="http://127.0.0.1:3002/"
+export STAGING_HOST_BYTE_LIMIT="549755813888000"
+
+cat <<ESQL | sqlite3 ./crates/banyan-core-service/data/server.db || fail 4 "creating staging server storage host record"
+INSERT INTO storage_hosts
+  (name, url, used_storage, available_storage, fingerprint, pem)
+  VALUES ('${STAGING_HOST_NAME}', '${STAGING_HOST_URL}', 0, ${STAGING_HOST_BYTE_LIMIT}, '${STAGING_HOST_FINGERPRINT}', '${STAGING_HOST_PUBKEY}');
+ESQL
 
 # Run storage provider
 (
@@ -48,22 +68,8 @@ cp -f crates/banyan-core-service/data/service-key.public crates/banyan-storage-p
 	timeout 3s cargo run || true
 )
 
-[ ! -f "crates/banyan-staging-service/data/service-key.public" ] && exit 8
-[ ! -f "crates/banyan-staging-service/data/service-key.fingerprint" ] && exit 9
-[ ! -f "crates/banyan-storage-provider-service/data/service-key.public" ] && exit 10
-[ ! -f "crates/banyan-storage-provider-service/data/service-key.fingerprint" ] && exit 11
-
-export STAGING_HOST_PUBKEY="$(cat crates/banyan-staging-service/data/service-key.public)"
-export STAGING_HOST_FINGERPRINT="$(cat crates/banyan-staging-service/data/service-key.fingerprint)"
-export STAGING_HOST_NAME="banyan-staging"
-export STAGING_HOST_URL="http://127.0.0.1:3002/"
-export STAGING_HOST_BYTE_LIMIT="549755813888000"
-
-cat <<ESQL | sqlite3 ./crates/banyan-core-service/data/server.db
-INSERT INTO storage_hosts
-  (name, url, used_storage, available_storage, fingerprint, pem)
-  VALUES ('${STAGING_HOST_NAME}', '${STAGING_HOST_URL}', 0, ${STAGING_HOST_BYTE_LIMIT}, '${STAGING_HOST_FINGERPRINT}', '${STAGING_HOST_PUBKEY}');
-ESQL
+[ -f "crates/banyan-storage-provider-service/data/service-key.public" ] || fail 5 "storage provider missing public service key"
+[ -f "crates/banyan-storage-provider-service/data/service-key.fingerprint" ] || fail 6 "storage provider missing service fingerprint"
 
 export STORAGE_HOST_PUBKEY="$(cat crates/banyan-storage-provider-service/data/service-key.public)"
 export STORAGE_HOST_FINGERPRINT="$(cat crates/banyan-storage-provider-service/data/service-key.fingerprint)"
@@ -71,7 +77,7 @@ export STORAGE_HOST_NAME="banyan-storage-provider"
 export STORAGE_HOST_URL="http://127.0.0.1:3003/"
 export STORAGE_HOST_BYTE_LIMIT="549755813888000"
 
-cat <<ESQL | sqlite3 ./crates/banyan-core-service/data/server.db
+cat <<ESQL | sqlite3 ./crates/banyan-core-service/data/server.db || fail 7 "creating storage server storage host record"
 INSERT INTO storage_hosts
   (name, url, used_storage, available_storage, fingerprint, pem)
   VALUES ('${STORAGE_HOST_NAME}', '${STORAGE_HOST_URL}', 0, ${STORAGE_HOST_BYTE_LIMIT}, '${STORAGE_HOST_FINGERPRINT}', '${STORAGE_HOST_PUBKEY}');
